moves generic components into temporary _generic directory
