+++ libs/menu-matriarch/ui/src/lib/_generic/card/card.component.scss
