+++ libs/menu-matriarch/ui/src/lib/_generic/checkbox/checkbox.component.scss
