+++ libs/menu-matriarch/ui/src/lib/_generic/options-menu/options-menu-animation.ts
