+++ libs/menu-matriarch/ui/src/lib/_generic/options-menu/options-menu-item/options-menu-item.component.scss
