+++ libs/menu-matriarch/ui/src/lib/_generic/small-caps-label/small-caps-label.component.scss
