+++ libs/menu-matriarch/ui/src/lib/_generic/tabset/tab/tab.component.ts
@@ -7,7 +7,7 @@ import {
   Output,
 } from '@angular/core';
 
-import { fadeInAnimation } from '../../animations';
+import { fadeInAnimation } from '../../../animations';
 
 @Component({
   selector: 'ui-tab',
